Refactor structure for donor and npo settings profile page
