+++ src/pages/settings-profile/donor/components/index.js
@@ -1,0 +1,1 @@
+export { default as DonorEditProfilePanel } from './DonorEditProfilePanel';
