+++ src/pages/settings-profile/npo/components/index.js
@@ -1,0 +1,1 @@
+export { default as NpoEditProfilePanel } from './NpoEditProfilePanel';
